+++ sdk/lib/async/timer.dart
@@ -33,6 +33,14 @@ abstract interface class Timer {
   ///
   /// The [callback] function is invoked after the given [duration].
   ///
+  /// The [callback] is registered in the current zone using
+  /// [Zone.registerCallback], and the timer is created using
+  /// that zone's [Zone.createTimer].
+  /// The default timer implementation will run the registered callback
+  /// in that zone using [Zone.run] when the timer expires,
+  /// and if the callback throws, errors will be reported as an
+  /// uncaught asynchronous error using the zone's [Zone.handleUncaughtError].
+  ///
   /// Example:
   /// ```dart
   /// final timer =
@@ -40,15 +48,11 @@ abstract interface class Timer {
   /// // Outputs after 5 seconds: "Timer finished".
   /// ```
   factory Timer(Duration duration, void Function() callback) {
-    if (Zone.current == Zone.root) {
-      // No need to bind the callback. We know that the root's timer will
-      // be invoked in the root zone.
-      return Zone.current.createTimer(duration, callback);
+    var zone = Zone._current;
+    if (!identical(zone, _rootZone)) {
+      callback = zone.registerCallback(callback);
     }
-    return Zone.current.createTimer(
-      duration,
-      Zone.current.bindCallbackGuarded(callback),
-    );
+    return zone.createTimer(duration, callback);
   }
 
   /// Creates a new repeating timer.
@@ -56,8 +60,19 @@ abstract interface class Timer {
   /// The [callback] is invoked repeatedly with [duration] intervals until
   /// canceled with the [cancel] function.
   ///
+  /// The [callback] is registered in the current zone using
+  /// [Zone.registerUnaryCallback], and the timer is created using
+  /// that zone's [Zone.createPeriodicTimer].
+  /// The default timer implementation will run the registered callback
+  /// in that zone using [Zone.runUnary] when the timer ticks,
+  /// and if the callback throws, errors will be reported as an
+  /// uncaught asynchronous error using the zone's [Zone.handleUncaughtError].
+  ///
+  /// If the callback throws, that will not cancel the timer.
+  ///
   /// The exact timing depends on the underlying timer implementation.
   /// No more than `n` callbacks will be made in `duration * n` time,
+  /// within the precision of the timer's granularity,
   /// but the time between two consecutive callbacks
   /// can be shorter and longer than `duration`.
   ///
@@ -86,25 +101,23 @@ abstract interface class Timer {
   /// // "Cancel timer"
   /// ```
   factory Timer.periodic(Duration duration, void callback(Timer timer)) {
-    if (Zone.current == Zone.root) {
-      // No need to bind the callback. We know that the root's timer will
-      // be invoked in the root zone.
-      return Zone.current.createPeriodicTimer(duration, callback);
+    Zone zone = Zone._current;
+    if (!identical(zone, _rootZone)) {
+      callback = zone.registerUnaryCallback(callback);
     }
-    var boundCallback = Zone.current.bindUnaryCallbackGuarded<Timer>(callback);
-    return Zone.current.createPeriodicTimer(duration, boundCallback);
+    return zone.createPeriodicTimer(duration, callback);
   }
 
   /// Runs the given [callback] asynchronously as soon as possible.
   ///
-  /// This function is equivalent to `new Timer(Duration.zero, callback)`.
+  /// This function is equivalent to `Timer(Duration.zero, callback)`.
   ///
   /// Example:
   /// ```dart
   /// Timer.run(() => print('timer run'));
   /// ```
   static void run(void Function() callback) {
-    new Timer(Duration.zero, callback);
+    Timer(Duration.zero, callback);
   }
 
   /// Cancels the timer.